# Move Rolestore.GetByNames query to raw query