@ethervoid 2020-04-28
@mgdelacroix I've refactored the query to use squirrel