@ethervoid 2020-04-29
Here you can see the improvements of this change:

```
benchmark                            old ns/op     new ns/op     delta
BenchmarkGetRoleNamesMySQL           292055        186175        -36.25%
BenchmarkGetRoleNamesPostgres        275924        179075        -35.10%
BenchmarkGetRoleNamesMySQL-2         296437        187681        -36.69%
BenchmarkGetRoleNamesPostgres-2      277950        180405        -35.09%
BenchmarkGetRoleNamesMySQL-4         295221        182264        -38.26%
BenchmarkGetRoleNamesPostgres-4      282121        172718        -38.78%
BenchmarkGetRoleNamesMySQL-8         305540        189448        -38.00%
BenchmarkGetRoleNamesPostgres-8      295626        164955        -44.20%
BenchmarkGetRoleNamesMySQL-16        305879        186824        -38.92%
BenchmarkGetRoleNamesPostgres-16     289715        195749        -32.43%

benchmark                            old allocs     new allocs     delta
BenchmarkGetRoleNamesMySQL           456            132            -71.05%
BenchmarkGetRoleNamesPostgres        455            147            -67.69%
BenchmarkGetRoleNamesMySQL-2         456            132            -71.05%
BenchmarkGetRoleNamesPostgres-2      455            147            -67.69%
BenchmarkGetRoleNamesMySQL-4         456            132            -71.05%
BenchmarkGetRoleNamesPostgres-4      455            147            -67.69%
BenchmarkGetRoleNamesMySQL-8         456            132            -71.05%
BenchmarkGetRoleNamesPostgres-8      455            147            -67.69%
BenchmarkGetRoleNamesMySQL-16        456            132            -71.05%
BenchmarkGetRoleNamesPostgres-16     455            147            -67.69%

benchmark                            old bytes     new bytes     delta
BenchmarkGetRoleNamesMySQL           12616         9296          -26.32%
BenchmarkGetRoleNamesPostgres        13904         11384         -18.12%
BenchmarkGetRoleNamesMySQL-2         12623         9296          -26.36%
BenchmarkGetRoleNamesPostgres-2      13909         11384         -18.15%
BenchmarkGetRoleNamesMySQL-4         12624         9296          -26.36%
BenchmarkGetRoleNamesPostgres-4      13911         11384         -18.17%
BenchmarkGetRoleNamesMySQL-8         12752         9297          -27.09%
BenchmarkGetRoleNamesPostgres-8      13992         11385         -18.63%
BenchmarkGetRoleNamesMySQL-16        12820         9297          -27.48%
BenchmarkGetRoleNamesPostgres-16     14072         11385         -19.09%
```

Here is the benchmark I made for it:

```go
func BenchmarkGetRoleNamesMySQL(b *testing.B) {
	b.StopTimer()
	st := &storeType{
		Name:        "MySQL",
		SqlSettings: storetest.MakeSqlSettings(model.DATABASE_DRIVER_MYSQL),
	}
	sqlStore := NewSqlSupplier(*st.SqlSettings, nil)
	s := &SqlRoleStore{sqlStore}
	createRoles(sqlStore)
	b.StartTimer()
	for n := 0; n < b.N; n++ {
		s.GetByNames([]string{"system_admin", "channel_admin", "channel_user"})
	}
}

func BenchmarkGetRoleNamesPostgres(b *testing.B) {
	b.StopTimer()
	st := &storeType{
		Name:        "Postgres",
		SqlSettings: storetest.MakeSqlSettings(model.DATABASE_DRIVER_POSTGRES),
	}
	sqlStore := NewSqlSupplier(*st.SqlSettings, nil)
	s := &SqlRoleStore{sqlStore}
	createRoles(sqlStore)
	b.StartTimer()
	for n := 0; n < b.N; n++ {
		s.GetByNames([]string{"system_admin", "channel_admin", "channel_user"})
	}
}

func createRoles(ss store.Store) {
	r1 := &model.Role{
		Name:        "channel_admin",
		DisplayName: model.NewId(),
		Description: model.NewId(),
		Permissions: []string{
			"add_reaction",
			"remove_reaction",
			"manage_public_channel_members",
			"manage_private_channel_members",
			"use_group_mentions",
			"use_channel_mentions",
			"manage_channel_roles create_post",
		},
		SchemeManaged: true,
	}
	ss.Role().Save(r1)
	r2 := &model.Role{
		Name:        "channel_user",
		DisplayName: model.NewId(),
		Description: model.NewId(),
		Permissions: []string{
			"remove_reaction",
			"create_post",
			"use_group_mentions",
			"upload_file",
			"edit_post",
			"use_channel_mentions",
			"delete_public_channel",
			"manage_private_channel_properties",
			"read_channel",
			"delete_post",
			"manage_public_channel_members",
			"manage_public_channel_properties",
			"manage_private_channel_members",
			"delete_private_channel",
			"get_public_link",
			"use_slash_commands",
			"add_reaction",
		},
		SchemeManaged: true,
	}
	ss.Role().Save(r2)
	r3 := &model.Role{
		Name:        "system_admin",
		DisplayName: model.NewId(),
		Description: model.NewId(),
		Permissions: []string{
			"edit_post", "create_post_public", "delete_private_channel", "invite_guest", "create_public_channel", "list_users_without_team", "create_private_channel", "delete_public_channel", "manage_jobs", "import_team", "create_emojis", "list_team_channels", "join_public_channels", "delete_others_posts", "remove_user_from_team", "manage_system_wide_oauth", "manage_private_channel_members", "manage_oauth", "create_bot", "manage_private_channel_properties", "promote_guest", "remove_others_reactions", "use_group_mentions", "edit_other_users", "create_post", "create_team", "manage_roles", "edit_others_posts", "manage_team", "manage_public_channel_members", "manage_others_outgoing_webhooks", "get_public_link", "list_private_teams", "manage_others_incoming_webhooks", "manage_outgoing_webhooks", "manage_incoming_webhooks", "delete_post", "read_others_bots", "revoke_user_access_token", "create_user_access_token", "use_channel_mentions", "read_user_access_token", "read_public_channel", "manage_bots", "invite_user", "view_team", "use_slash_commands", "manage_team_roles", "add_user_to_team", "view_members", "manage_others_slash_commands", "upload_file", "manage_others_bots", "manage_system", "add_reaction", "delete_emojis", "assign_system_admin_role", "manage_public_channel_properties", "demote_to_guest", "remove_reaction", "manage_slash_commands", "manage_channel_roles", "delete_others_emojis", "read_bots", "create_post_ephemeral", "read_channel", "join_private_teams",
		},
		SchemeManaged: true,
	}
	ss.Role().Save(r3)
}
```