@agnivade 2020-04-29
Great, thanks ! A microbenchmark is not super-ideal to test things happening outside the process boundary as there's network latency involved. 

If you don't mind, I'll give this a spin with the new load tests and report back results. That should give a more realistic estimate.