@ethervoid 2020-04-29
Well in its pure form, letting aside the network latency, we can see a good improvement for ops as well as for memory.

I'm eager to see how it behaves with the load tester :D