@agnivade 2020-04-29
Yep. This is more to get some high-level numbers. This is what was done for other PRs too.

Micro benchmarks require the numbers to be extremely stable with a variance of <5%. And your DB was installed locally, the numbers would vary more widely if the DB was installed on a separate machine.