@agnivade 2020-04-30
Nvm, this is a cached query. So hard to measure improvements this way. Microbenchmarks will have to do for now.